Fix preprocessor if wrong precedence ordering when using macros Refractor for the glsl parser Partial primary expression parser
